refactor(api): move user controllers
to user directory
