+++ api/src/identity-access-management/application/user/user.admin.route.js
@@ -1,7 +1,7 @@
 import Joi from 'joi';
 
-import { securityPreHandlers } from '../../shared/application/security-pre-handlers.js';
-import { identifiersType } from '../../shared/domain/types/identifiers-type.js';
+import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
+import { identifiersType } from '../../../shared/domain/types/identifiers-type.js';
 import { userAdminController } from './user.admin.controller.js';
 
 export const userAdminRoutes = [
