+++ api/tests/identity-access-management/unit/application/user.controller.test.js
@@ -1,5 +1,5 @@
 import * as requestResponseUtils from '../../../../lib/infrastructure/utils/request-response-utils.js';
-import { userController } from '../../../../src/identity-access-management/application/user.controller.js';
+import { userController } from '../../../../src/identity-access-management/application/user/user.controller.js';
 import { usecases } from '../../../../src/identity-access-management/domain/usecases/index.js';
 import { User } from '../../../../src/shared/domain/models/User.js';
 import { expect, hFake, sinon } from '../../../test-helper.js';
